orientation-event/DeviceOrientation/angle-set-null.html

<!DOCTYPE html>
<html>

  <head>
    <title>Unable to provide angles must set values to null:</title>
 	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
   <script src="http://w3c-test.org/resources/testharness.js"></script>
    <script src="http://w3c-test.org/resources/testharnessreport.js"></script>
    <link rel='author' title='Mosquito FP7'/>
  </head>

  <body>

  
  		<h1>Unable to provide angles must set values to null:</h1>
    <div id="log">
    </div>
	
	<script>
	absolute = null;
	   if (window.DeviceOrientationEvent) {
	  		window.addEventListener('deviceorientation', function(eventData) {
	
				// Gamma : angle par rapport a x
				gamma = eventData.gamma;
				// Beta : angle par rapport a y
				beta = eventData.beta;
				// Alpha : orientation (N-S-E-O)
				alpha = eventData.alpha;
				absolute = eventData.absolute;
				
			}, false);
			
		}
	
	alert("Launch Test.");
	alert(absolute);
	test(function() {assert_equals(gamma, null)}, "gamma must be to null");
	test(function() {assert_equals(beta, null)}, "beta must be to null");
	test(function() {assert_equals(alpha, null)}, "alpha must be to null");
	  
	</script>
	
  </body>
  
</html>
